+++ BigScreen/Pages/TifoAdm/GetVideoSize.cshtml.cs
@@ -84,8 +84,8 @@ namespace BigScreen
 
 
 
-                        Tifo.TifoHeight = (int)(Convert.ToDecimal(obj.VideoHeight));
-                        Tifo.TifoWidth = (int)(Convert.ToDecimal(obj.VideoWidth));
+                        Tifo.TifoHeight = (int)(Convert.ToInt32(obj.VideoHeight));
+                        Tifo.TifoWidth = (int)(Convert.ToInt32(obj.VideoWidth));
                         Tifo.Id = Convert.ToInt32(obj.Id);
                         Tifo.Title = obj.Title;
                         Tifo.Path = obj.Path;
